Crate gomori

Overview

A bot will receive a Board and a set of five Cards, and outputs a list of CardToPlay. Therefore, to get familiar with the API, it is recommended to start looking at these types.

Coordinates

Gomori uses 2D coordinates, usually referred to as i and j, to identify locations on the board. i is always the first coordinate in a coordinate pair, and j the second.

Interpretation

This library doesn’t define which of i and j is the horizontal coordinate and which is the vertical, or if larger coordinates mean “left” or “right”, “down” or “up”. The game rules can be implemented without assigning an interpretation to these coordinates, and so it is up to the user to choose how to interpret and visualize these coordinates. For instance, one could treat them as matrix coordinates, i.e. i being the row index (top to bottom) and j the column index (left to right).

Absolute vs relative

The game rules allow the board location to drift over the course of a game. For example, imagine that players play new cards (darker shading) to the right of the board, and winning (taking away) cards on the left like so:

▒ ▒ ▒ ▓          ▒          ▒          ▒
  ▒ ▒ ▒  =>  ▒ ▒ ▒  =>  ▒ ▒ ▒ ▓  =>      ▒
    ▒ ▒        ▒ ▒        ▒ ▒        ▒ ▒

In each turn, the board fits within a 4x4 grid, but the 4x4 grid shifts to the right.

This library uses absolute coordinates, meaning that the same “spot on the table”, called a “field” by this library, has consistent coordinates in every turn, no matter where the current 4x4 boundary is. As a result, these coordinates may be negative, or larger than 4. They are represented as an i8.

Macros

card

Shorthand for creating cards from a two-character string.

Structs

BitBoard

A compact board representation that stores only a single bit per field, equivalent to a set of coordinates.

BitBoardIter

Iterator produced by BitBoard::into_iter().

Board

Represents a board with at least one card on it.

BoundingBox

A 2D area represented by a min + max coordinate pair.

CalculatedEffects

The effects that playing a card would have.

Card

A playing card in a standard 52-card game.

CardToPlay

Specifies which card to play, and where.

CardsSet

A compact set of Cards.

CardsSetIter

Iterator for a CardsSet that returns cards by ascending rank.

CompactField

A compact representation of a single field on the board.

Field

A single field on the board, including coordinates.

IllegalCardPlayedException

Describes why the card cannot be played.

IllegalMoveException

Describes why a move is illegal.

Okay

Dummy struct for use in bot communication.

PlayTurnResponse

The cards to play in this turn, in order.

PlayerState

The state for a single player during one game.

Enums

CardFromStrErr

The error type for the FromStr instance of Card.

Color

Black or white.

IllegalCardPlayed

The error type for Board::calculate(), i.e. for playing a single card.

IllegalMove

The error type for one turn.

Rank

The rank of a card.

Request

Request for a bot to do something.

Suit

The suit of a card.

TurnOutcome

Summarizes the outcome of playing a turn (i.e. playing up to five cards).

Constants

BLACK_CARDS_SET

Equal to CardsSet::from_iter(BLACK_CARDS).

BOARD_SIZE

RED_CARDS_SET

Equal to CardsSet::from_iter(RED_CARDS).

Statics

BLACK_CARDS

RED_CARDS

Functions

execute_first_turn

execute_turn

visualize_top_cards