Struct gomori::BitBoard

pub struct BitBoard { /* private fields */ }

A compact board representation that stores only a single bit per field, equivalent to a set of coordinates.

Every BitBoard is obtained from a Board, and can, for instance, store which fields on the board have a visible diamond card on them. So a BitBoard is kind of like a “view” into a particular Board.

Two BitBoards can be combined with operations like &, |, or difference(). It’s important that when two BitBoards are combined, they must both originate1 from the same Board. The reason is that the area represented by BitBoards from different Boards may not be identical, so some contents of one board may be unrepresentable in the other BitBoard. This is only checked with a debug assertion.

It can be converted back into a list of coordinate pairs by means of its IntoIterator instance.

Note on immutability

This is an immutable type, so its “mutating” methods return a new value instead of really mutating in-place (except for std::ops::BitXxxAssign trait methods). It is also Copy, so a value is not consumed by methods with self receiver.

Implementation

Internally, a BitBoard is (1) an (offset_i, offset_j) coordinate pair and (2) a 49-bit-bitset. The bitset encodes an 7 x 7 area with one bit per field, like so:

1 0 1 1 0 0 1
1 1 1 0 1 0 1
0 1 1 0 0 1 0
0 0 1 1 1 1 0
0 1 1 0 1 0 1
0 0 0 1 0 1 1
0 0 0 1 0 0 1

The offset is added to every local (i, j) coordinate in that field (the local i and j both range from 0 to 6) to obtain the true (i, j) coordinate.

Every valid board would fit in a 4 x 4 area, so why 7 x 7? One reason is that with a 7 x 7 board, we can be sure that not only the board itself can be represented, but also the next card, as long as it is in the board’s playable area.

“BitBoard X originates from Board B” means that that X was returned by some method of B, or that X is the result of combining/processing BitBoards that originate from X. ↩

Implementations

impl BitBoard

pub fn contains(self, i: i8, j: i8) -> bool

Checks whether the given coordinate is contained in the set.

pub fn insert(self, i: i8, j: i8) -> Self

Sets the bit for the specified coordinate to true.

This function must only be used with coordinates in the underlying board’s playable_area. Other coordinates may exceed the 7x7 area stored in the BitBoard, and that will cause a panic in debug mode. In release mode, no checks are performed, it will just cause invalid data.

pub fn remove(self, i: i8, j: i8) -> Self

Sets the bit for the specified coordinate to true.

The same restriction as for insert() applies.

pub fn is_empty(self) -> bool

pub fn num_entries(self) -> u32

pub fn difference(self, other: BitBoard) -> BitBoard

Computes the difference to another BitBoard.

pub fn lines_going_through_point(self, point_i: i8, point_j: i8) -> BitBoard

Checks whether there are any horizontal, vertical or diagonal lines of length 4 passing through the specified point (in a 7 x 7 area centered on the point).

Any lines that are found are returned in a new BitBoard. The result is therefore a subset of the input.

Only valid for point coordinates in the range [-52, 52].