Struct gomori::BoundingBox

pub struct BoundingBox {
    pub i_min: i8,
    pub j_min: i8,
    pub i_max: i8,
    pub j_max: i8,
}

A 2D area represented by a min + max coordinate pair.

The two coordinates form an inclusive 2D range, i.e. unlike in a half-open range, it’s possible for a point with i == i_max to be contained in the area.

Fields
i_min: i8j_min: i8i_max: i8j_max: i8
Implementations

impl BoundingBox

pub fn size_i(&self) -> u8

The size of the bounding box along the i dimension.

pub fn size_j(&self) -> u8

The size of the bounding box along the j dimension.

pub fn contains(&self, i: i8, j: i8) -> bool

pub fn singleton(i: i8, j: i8) -> Self

pub fn from_coordinates_iter( iter: impl Iterator<Item = (i8, i8)> ) -> Option<Self>

pub fn update(&mut self, i: i8, j: i8)

Expands the bounding box to cover point (i, j).