Struct gomori::CardsSet

pub struct CardsSet { /* private fields */ }

A compact set of Cards.

Allows intersection/union/xor with other such sets via bitwise ops. Also implements IntoIterator, so it can be converted into e.g. a vector with Vec::from_iter(cards_set).

use gomori::{card, CardsSet};
let mut set = CardsSet::new();
// This is an immutable data type, so functions like `insert` return a new `CardsSet`.
set = set.insert(card!("7♥"));
set = set.insert(card!("7♥"));  // Inserting a second time has no effect
set = set.insert(card!("2♥"));
assert_eq!(Vec::from_iter(set), vec![card!("2♥"), card!("7♥")]);

Note on immutability

This is an immutable type, so its “mutating” methods return a new value instead of really mutating in-place (except for std::ops::BitXxxAssign trait methods). It is also Copy, so a value is not consumed by methods with self receiver.

Implementations

impl CardsSet

pub fn new() -> Self

Creates a new, empty set.

pub fn len(self) -> u32

pub fn contains(self, card: Card) -> bool

pub fn is_empty(self) -> bool

pub fn insert(self, card: Card) -> Self

pub fn remove(self, card: Card) -> Self